backend/jtabackend/ml_scripts/requirements.txt
torch==2.0.1
transformers==4.46.0
numpy<2.0
python-Levenshtein
fuzzywuzzy
flask
flask_cors
spacy
PyPDF2
python-docx